backend/requirements.txt
Django
djangorestframework
django[argon2]
django-filter
djangorestframework-simplejwt
